+++ package.json
@@ -172,6 +172,7 @@
     "@protobuf-ts/protoc": "^2.8.0",
     "aegir": "^37.6.6",
     "it-first": "^2.0.0",
+    "it-pair": "^2.0.3",
     "libp2p": "^0.41.0",
     "protons": "^6.0.1"
   }
