+++ src/peer_transport/transport.ts
@@ -3,23 +3,17 @@ import { CreateListenerOptions, DialOptions, Listener, symbol, Transport } from 
 import type { ConnectionHandler, TransportManager, Upgrader } from '@libp2p/interface-transport'
 import { multiaddr, Multiaddr } from '@multiformats/multiaddr'
 import type { IncomingStreamData, Registrar } from '@libp2p/interface-registrar'
-import { pbStream } from 'it-pb-stream'
-import pDefer, { DeferredPromise } from 'p-defer'
 import type { PeerId } from '@libp2p/interface-peer-id'
-import { abortableDuplex } from 'abortable-iterator'
-import { TimeoutController } from 'timeout-abort-controller'
 import { WebRTCMultiaddrConnection } from '../maconn.js'
 import type { Startable } from '@libp2p/interfaces/startable'
 import { DataChannelMuxerFactory } from '../muxer.js'
 import { WebRTCPeerListener } from './listener.js'
 import type { PeerStore } from '@libp2p/interface-peer-store'
 import { logger } from '@libp2p/logger'
-import * as pb from './pb/index.js'
-import { readCandidatesUntilConnected } from './util.js'
+import { connect, handleIncomingStream } from './handler.js'
 
 const log = logger('libp2p:webrtc:peer')
 
-const TIMEOUT = 30 * 1000
 export const TRANSPORT = '/webrtc-direct'
 export const PROTOCOL = '/webrtc-direct/0.0.1'
 export const CODE = 276
@@ -87,11 +81,7 @@ export class WebRTCDirectTransport implements Transport, Startable {
    * <relay address>/p2p/<relay-peer>/p2p-circuit/webrtc-direct/p2p/<destination-peer>
   */
   async dial (ma: Multiaddr, options: DialOptions): Promise<Connection> {
-    // extract peer id
-    // if (!remotePeerId) {
-    //   throw("peerId should be present in multiaddr")
-    // }
-
+    log.trace('dialing address: ', ma)
     const addrs = ma.toString().split(TRANSPORT)
     // look for remote peerId
     let relayed = multiaddr(addrs[0])
@@ -99,59 +89,19 @@ export class WebRTCDirectTransport implements Transport, Startable {
     if (remotePeerId != null) {
       relayed = relayed.encapsulate(multiaddr(`/p2p/${remotePeerId}`))
     }
+    const controller = new AbortController()
     if (options.signal == null) {
-      options.signal = new AbortSignal()
+      options.signal = controller.signal
     }
 
     const connection = await this.components.transportManager.dial(relayed)
     const rawStream = await connection.newStream([PROTOCOL], options)
-    const stream = pbStream(abortableDuplex(rawStream, options.signal)).pb(pb.Message)
 
-    // setup peer connection
-    const pc = new RTCPeerConnection(this.init.rtcConfiguration)
-    // the label is not relevant to connection initiation but can be
-    // useful for debugging
-    const channel = pc.createDataChannel('init')
-
-    const connectedPromise = pDefer<void>()
-    pc.onconnectionstatechange = (_) => {
-      switch (pc.connectionState) {
-        case 'connected':
-          return connectedPromise.resolve()
-        case 'closed':
-        case 'disconnected':
-        case 'failed':
-          return connectedPromise.reject()
-      }
-    }
-
-    options.signal.onabort = connectedPromise.reject
-    // setup callback to write ICE candidates to the remote
-    // peer
-    pc.onicecandidate = ({ candidate }) => {
-      stream.write({
-        type: pb.Message.Type.ICE_CANDIDATE,
-        data: (candidate != null) ? JSON.stringify(candidate) : ''
-      })
-    }
-
-    // read offer
-    const offerMessage = await stream.read()
-    if (offerMessage.type !== pb.Message.Type.SDP_OFFER) {
-      throw new Error('remote should send an SDP offer')
-    }
-
-    const offerSdp = new RTCSessionDescription({ type: 'offer', sdp: offerMessage.data })
-    await pc.setRemoteDescription(offerSdp)
-
-    // create an answer
-    const answerSdp = await pc.createAnswer()
-    // write the answer to the stream
-    stream.write({ type: pb.Message.Type.SDP_ANSWER, data: answerSdp.sdp })
-    // set answer as local description
-    pc.setLocalDescription(answerSdp)
-
-    await readCandidatesUntilConnected(connectedPromise, pc, stream)
+    const pc = await connect({
+      stream: rawStream,
+      rtcConfiguration: this.init.rtcConfiguration,
+      signal: options.signal
+    })
 
     const result = options.upgrader.upgradeOutbound(
       new WebRTCMultiaddrConnection({
@@ -167,59 +117,18 @@ export class WebRTCDirectTransport implements Transport, Startable {
     )
     // close streams
     rawStream.close()
-    channel.close()
     void connection.close()
     // TODO: hack
     // await new Promise(res => setTimeout(res, 100))
     return await result
   }
 
-  async _onProtocol ({ connection, stream: rawStream }: IncomingStreamData) {
-    const timeoutController = new TimeoutController(TIMEOUT)
-    const signal = timeoutController.signal
-    const stream = pbStream(abortableDuplex(rawStream, timeoutController.signal)).pb(pb.Message)
-    const pc = new RTCPeerConnection(this.init.rtcConfiguration)
-
-    const connectedPromise: DeferredPromise<void> = pDefer()
-    signal.onabort = () => connectedPromise.reject()
-    // candidate callbacks
-    pc.onicecandidate = ({ candidate }) => {
-      stream.write({
-        type: pb.Message.Type.ICE_CANDIDATE,
-        data: (candidate != null) ? JSON.stringify(candidate) : ''
-      })
-    }
-    pc.onconnectionstatechange = (_) => {
-      log.trace('received pc state: ', pc.connectionState)
-      switch (pc.connectionState) {
-        case 'connected':
-          connectedPromise.resolve()
-          break
-        case 'failed':
-        case 'disconnected':
-        case 'closed':
-          connectedPromise.reject()
-      }
-    }
-
-    // create and write an SDP offer
-    const offer = await pc.createOffer()
-    pc.setLocalDescription(offer)
-    stream.write({ type: pb.Message.Type.SDP_OFFER, data: offer.sdp })
-
-    // read an SDP anwer
-    const pbAnswer = await stream.read()
-    if (pbAnswer.type != pb.Message.Type.SDP_ANSWER) {
-      throw new Error('response message should be an SDP answer')
-    }
-    const answer = new RTCSessionDescription({
-      type: 'answer',
-      sdp: pbAnswer.data
+  async _onProtocol ({ connection, stream }: IncomingStreamData) {
+    const pc = await handleIncomingStream({
+      rtcConfiguration: this.init.rtcConfiguration,
+      connection,
+      stream
     })
-    await pc.setRemoteDescription(answer)
-
-    // wait until candidates are connected
-    await readCandidatesUntilConnected(connectedPromise, pc, stream)
     const muxerFactory = new DataChannelMuxerFactory(pc, '/webrtc-direct')
     const conn = await this.components.upgrader.upgradeInbound(new WebRTCMultiaddrConnection({
       peerConnection: pc,
@@ -230,7 +139,7 @@ export class WebRTCDirectTransport implements Transport, Startable {
       skipProtection: true,
       muxerFactory
     })
-    rawStream.close()
+    stream.close()
     if (this.handler != null) {
       this.handler(conn)
     }
